Try changing Matrix for Array in some math functions
Array is more appropriate if we're thinking in the context of collecting
numbers to do vectorized operations rather than literal matrices that
represent linear operators and elements of a Hilbert space.

It's possible this will break some things, and it's also possible this
change will need to be propagated further up the call tree.  Let's see
how this goes.