Fix the gradient iteration order in SphericalExpansion
(and in the finite-difference test).  They're now iterating in the
order we originally intended -- grad[neigh] stores the gradient of the
_center_ coefficients w.r.t. the motion of the _neighbour_ atom.

The finite-difference test, however, still moves only the center (for
simplicity and efficiency), so we have to account for this.  Each
neighbour gradient contribution only contributes to one key (species),
which has been fixed in both f() and grad_f().  But we don't yet have a
way of indexing grad[swap(neigh)], which we need to get a consistent
grad_f().