Couple the RepManagerGradientFixture more closely with the Calculator
They refer to the same thing and their behaviour is also closely coupled
(they must iterate over the list of centers in lockstep; the class has
been modified to enforce this behaviour).

The implementation of the above behaviour as a C++ iterator is left as an
exercise to the reader.