Add optional fixed-size argument vectors for gradient test
This mode is activated by defining a const static size_t named
`n_arguments` in the calculator class; the functions `f()` and
`grad_f()` can then be made to accept fixed-size vectors as input.
Note also that the Jacobian output by `grad_f()` is expected to
have as many rows as the length of the argument vector, consistent with
the usual definition of the Jacobian matrix.

This fix is motivated by compiler complaints in some versions of Eigen,
which didn't like passing a dynamic-sized vector to a function taking
fixed-size vectors.  For calculators whose argument size is known at
compile time (e.g. spherical harmonics or spherical expansion), the size
checking can now be done entirely at compile time.