Re-enable breathe for doxygen integration
Turns out it wasn't working because the directives were just commented
out :facepalm:.  Anyway, it works now (with a shitload of warnings, but
at least it builds now).  It's going to take quite a lot of human effort
to organize the documentation (sadly it's not done for you, like in
Doxygen) but hey, at least LaTeX maths works.