Essence of a fix for the tests (works on some structures)
This fixes the swap_pair_ref function to pick the i atom (old center,
new neighbour) based on position, not index, in case it appears as a
periodic image in j's neighbourhood.

But this doesn't fix it properly, because there might be multiple images
of i in j's neighbourhood, and they all contribute to the gradient with
respect to movement of i.