Bump up the gradient error tolerance (again) for small_molecule.json
Apparently the SOAP finite-difference gradient for this molecule never
quite reaches the previous limit, due to some adverse interaction
between high truncation error and weirdly large numerical error (perhaps
this needs to be investigated? -- run it with verbosity=DEBUG to see).

In any case, we might want a way to override the default
finite-difference error tolerance per structure and representation.