Change Eigen Refs to proper arrays in Hyp1f1GradientProvider
The functions f() and grad_f() didn't have any persistent storage of the
Eigen Maps they were returning (since those in turn referenced data
stored in the Hyp1f1Calculator), so returning Refs to these objects was
evidently a bad idea -- even if it did seem to work for all the cases we
tested.  Oh well.