Start a k-space version of calculator_spherical_expansion.hh
This is based off the (real-space) version in the aforementioned file,
but with AtomicSmearingType mostly stripped out (we don't need it; just
use a constant smearing).  The class has been renamed but I still don't
know what to do with the radial contribution class -- rename that as
well, or add a template parameter?

Notes have been added (with the text MV NOTES) to places where
implementation will need to be added / changed significantly for k-space.