Fixed a conceptual bug in the calculation of spherical covariants
A one-year old "tidy-up" commit changed the order of evaluation of the
loop that fills the wigner3j precomputer. This messes up the calculation
of the covariants, that expect the same loop order, and mysteriously did
so in a way that preserves the covariant nature of the output.

This should however be at least consistent, and should do until we have
a cleaner, more comprehensive implementation of equivariants