Fix gradients with optimal_basis GTOs (#365)
Bugfix for optimal radial basis

finalize_gradients function was not overridden properly. It was not caught by
the tests because the gradient was only tested for compute_neighbour_derivative
but not the whole computation of the sph exp coeffs, such tests have been added
now to test_calculator/spherical_representation_gradients also for DVR and for
Spline separately. Testing time of test_calculator increased from 8 to 12 s

contains also some small fixes related to utilities for computation of optimal
radial basis